requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1673535674859/work
addict @ file:///home/conda/feedstock_root/build_artifacts/addict_1636818143388/work
aiofiles==22.1.0
aiosqlite==0.19.0
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
bleach==6.0.0
blessings @ file:///tmp/build/80754af9/blessings_1614076441300/work
blinker==1.6.2
blis==0.7.9
braceexpand==0.1.7
brotlipy==0.7.0
cachetools==5.3.0
catalogue==2.0.8
ccimport==0.4.2
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
cfgv==3.3.1
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm==0.1.3
confection==0.0.4
contexttimer==0.3.3
contourpy==1.0.7
cryptography @ file:///croot/cryptography_1677533068310/work
cumm-cu113==0.4.8
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.7
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.6
einops @ file:///home/conda/feedstock_root/build_artifacts/einops_1670600230829/work
entrypoints==0.4
executing==1.2.0
fairscale==0.4.4
fastjsonschema==2.16.3
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1680834529360/work
fire==0.5.0
Flask==2.2.3
Flask-API==3.0.post1
flit_core @ file:///croot/flit-core_1679397103445/work/source/flit_core
fonttools==4.39.3
fqdn==1.5.1
ftfy==6.1.1
gcc7==0.0.7
gitdb==4.0.10
GitPython==3.1.31
gpustat @ file:///tmp/build/80754af9/gpustat_1621432778913/work
grpcio @ file:///tmp/build/80754af9/grpcio_1637590823556/work
h5py @ file:///tmp/abs_4aewd3wzey/croots/recipe/h5py_1659091371897/work
huggingface-hub @ file:///home/conda/feedstock_root/build_artifacts/huggingface_hub_1681153468624/work
identify==2.5.22
idna @ file:///croot/idna_1666125576474/work
imageio==2.27.0
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1681581693083/work
importlib-resources==5.12.0
iopath==0.1.10
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.4
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_packaging==0.12.3
jupyter_server==2.5.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.22.1
kaggle==1.5.13
kiwisolver==1.4.4
langcodes==3.3.0
lark==1.1.5
lazy_loader==0.2
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1679584000376/work
markdown-it-py==2.2.0
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1674135809652/work
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
murmurhash==1.0.9
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
ninja==1.11.1
nodeenv==1.7.0
notebook==6.5.4
notebook_shim==0.2.2
numpy @ file:///croot/numpy_and_numpy_base_1672336185480/work
nvidia-ml-py3 @ file:///tmp/build/80754af9/nvidia-ml_1621428042568/work
omegaconf==2.3.0
open3d-python==0.2.0.0
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.1
pccm==0.4.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
plotly==5.14.1
plyfile @ file:///home/conda/feedstock_root/build_artifacts/plyfile_1679423307575/work
pointops==1.0
pointops2==1.0
portalocker==2.7.0
pre-commit==3.2.2
preshed==3.0.8
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.15.8
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pybind11==2.10.4
pycocoevalcap==1.2
pycocotools==2.0.6
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.0
Pympler==1.0.1
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
python-json-logger==2.0.7
python-magic==0.4.27
python-slugify==8.0.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML @ file:///croot/pyyaml_1670514731622/work
pyzmq==25.0.2
regex==2023.3.23
requests @ file:///croot/requests_1678709721434/work
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.4
scikit-image==0.20.0
scikit-learn==1.2.2
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy_1653073867187/work
Send2Trash==1.8.0
sentencepiece==0.1.98
SharedArray @ file:///home/conda/feedstock_root/build_artifacts/sharedarray_1667184240327/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spconv-cu113==2.3.5
srsly==2.4.6
stack-data==0.6.2
streamlit==1.21.0
tenacity==8.2.2
tensorboard==1.15.0
tensorboardX @ file:///home/conda/feedstock_root/build_artifacts/tensorboardx_1654638215170/work
termcolor @ file:///home/conda/feedstock_root/build_artifacts/termcolor_1672833821273/work
terminado==0.17.1
text-unidecode==1.3
thinc==8.1.9
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.4.12
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.7
toolz==0.12.0
torch==1.12.1
torch-cluster @ file:///usr/share/miniconda/envs/test/conda-bld/pytorch-cluster_1656604202270/work
torch-geometric==2.3.0
torch-scatter @ file:///usr/share/miniconda/envs/test/conda-bld/pytorch-scatter_1669298754121/work
torch-sparse @ file:///usr/share/miniconda/envs/test/conda-bld/pytorch-sparse_1671714926258/work
torchaudio==0.12.1
torchvision==0.13.1
tornado==6.2
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1677948868469/work
traitlets==5.9.0
transformers==4.28.1
typer==0.7.0
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
tzdata==2023.3
tzlocal==4.3
uri-template==1.2.0
urllib3 @ file:///croot/urllib3_1680254681959/work
validators==0.20.0
virtualenv==20.21.0
wasabi==1.1.1
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webdataset==0.2.48
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1676411946679/work
widgetsnbextension==4.0.7
y-py==0.5.9
yapf @ file:///home/conda/feedstock_root/build_artifacts/yapf_1641487982943/work
ypy-websocket==0.8.2
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1677313463193/work
